+++ nekoyume/Assets/MOD/Ares/package.json
@@ -1,7 +1,7 @@
 {
     "name": "com.planetariumlabs.9c-unity-mod-ares",
     "version": "0.1.0",
-    "displayName": "Ares",
+    "displayName": "NineChronicles Mod Ares",
     "description": "This is a mod project used by the Unity client for Nine Chronicles.",
     "unity": "2021.3",
     "dependencies": {},
@@ -10,8 +10,7 @@
         "ninechronicles",
         "9c",
         "mod",
-        "ares",
-        "arena"
+        "ares"
     ],
     "license": "AGPL-3.0"
 }
