+++ nekoyume/Assets/MOD/Core.meta
@@ -1,0 +1,3 @@
+fileFormatVersion: 2
+guid: aa2f8bcaee684647b44b4bac4363fe77
+timeCreated: 1707280838
