+++ nekoyume/Assets/MOD/Core/package.json
@@ -1,0 +1,16 @@
+{
+  "name": "com.planetariumlabs.9c-unity-mod-core",
+  "version": "0.1.0",
+  "displayName": "NineChronicles Mod Core",
+  "description": "This is a mod project used by the Unity client for Nine Chronicles.",
+  "unity": "2021.3",
+  "dependencies": {},
+  "keywords": [
+    "planetariumlabs",
+    "ninechronicles",
+    "9c",
+    "mod",
+    "core"
+  ],
+  "license": "AGPL-3.0"
+}
