+++ nekoyume/Assets/MOD/Core/package.json.meta
@@ -1,0 +1,3 @@
+fileFormatVersion: 2
+guid: fc599152ad6947df8123e275ac2f0885
+timeCreated: 1707281153
