+++ nekoyume/Assets/MOD/README.md.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: fc75dacfa55c6434e91fa327be698333
+TextScriptImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
